add javadocs and fix codestyle
