--- src/main/java/Cleese/InvalidInputException.java
@@ -1,7 +1,0 @@
-package Cleese;
-
-public class InvalidInputException extends Exception{
-    InvalidInputException(String errorMessage) {
-        super(errorMessage);
-    }
-}
