+++ src/main/java/exceptions/NoDescException.java
@@ -1,5 +1,8 @@
-package Cleese;
+package exceptions;
 
+/**
+ * Exception to represent the case when there is no description given
+ */
 public class NoDescException extends Exception{
     public NoDescException(String errorMessage) {
         super(errorMessage);
