+++ src/test/java/CleeseTest/CleeseTest.java
@@ -1,7 +1,8 @@
 package CleeseTest;
 
-import Cleese.*;
+import cleese.*;
 
+import task.TasksList;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -9,13 +10,13 @@ import static org.junit.jupiter.api.Assertions.fail;
 public class CleeseTest {
     private Parser parser = new Parser();
     private Storage storage = new Storage("placeholder");
-    private TaskList taskList = new TaskList();
+    private TasksList tasksList = new TasksList();
     private Ui ui = new Ui();
 
     @Test
     public void testListCommand() {
         try {
-            assertEquals(parser.handleCommand("list",taskList,ui,storage),"list" );
+            assertEquals(parser.handleCommand("list", tasksList,ui,storage),"list" );
         } catch (Exception e) {
             fail();
         }
@@ -23,7 +24,7 @@ public class CleeseTest {
     @Test
     public void testByeCommand() {
         try {
-            assertEquals(parser.handleCommand("bye",taskList,ui,storage),"bye" );
+            assertEquals(parser.handleCommand("bye", tasksList,ui,storage),"bye" );
         } catch (Exception e) {
             fail();
         }
